requirements.txt
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
livekit-plugins-deepgram>=0.6.17,<1.0.0
livekit-plugins-silero>=0.7.4,<1.0.0
livekit-plugins-turn-detector>=0.4.0,<1.0.0
python-dotenv~=1.0